Add macros.ea-apu
Case ZC-839:  More cascading configurations, and also using some of the
macros we set up in ea-apr.

Additionally, some subpackages (mod_suphp) get very upset when the
apr-util header files are not in the same directory as the apr header
files, so we needed to move this into the same directory tree as apr.